Common adjustments

This family of adj_*() functions encode business logic for common adjustments made in the financial world.

adj_following()

Choose the first non-event date after x. This is equivalent to using an adjustment of +days(1).

adj_preceding()

Choose the first non-event date before x. This is equivalent to using an adjustment of -days(1).

adj_modified_following()

Choose the first non-event date after x, unless it falls in a different month, in which case the first non-event date before x is chosen instead.

adj_modified_preceding()

Choose the first non-event date before x, unless it falls in a different month, in which case the first non-event date after x is chosen instead.

adj_nearest()

Choose the nearest non-event date to x. If the closest preceding and following non-event dates are equally far away, the following non-event date is chosen.

adj_following(x, schedule)

adj_preceding(x, schedule)

adj_modified_following(x, schedule)

adj_modified_preceding(x, schedule)

adj_nearest(x, schedule)

Arguments

x
[Date]

A vector of dates.

schedule
[schedule / event]

A schedule or event.

Details

For examples, see sch_adjust().