Create a business day period

bdays() is similar to lubridate::days() in that it constructs an object that can be added to Dates to shift by a set period. It differs in the fact that you can provide an rbundle to shift relative to.

bdays(x, rbundle)

Arguments

x
[integer]

The number of days to step. Can be negative to step backwards.

rbundle
[rbundle / rrule]

An rbundle or rrule.

Details

When added to a Date, this internally calls alma_step(), meaning that an adjustment is performed after every 1 day step. This is likely the behavior that is desired for shifting by "n business days".

Examples

# Create a recurrence rule for weekends
on_weekends <- weekly() %>%
  recur_on_weekends()

# A Thursday and Friday
x <- as.Date(c("1970-01-01", "1970-01-02"))

# lubridate::days() just shifts forward by 2 days, without paying
# attention to the weekend
x + days(2)

#> [1] "1970-01-03" "1970-01-04"

# With bdays(), you can specify an rbundle to step relative to. This steps
# over the weekend, landing us on Monday and Tuesday.
x + bdays(2, on_weekends)

#> [1] "1970-01-05" "1970-01-06"

# It is also vectorized, and arithmetic is defined on integer input
business_days <- bdays(2, on_weekends) * c(2, 3)
business_days

#> <bdays<1 rrules / 0 rdates / 0 exdates>>
#> [1] 4 6

x + business_days

#> [1] "1970-01-07" "1970-01-12"

x - business_days

#> [1] "1969-12-28" "1969-12-27"