Shift dates relative to a business calendar

cal_shift() shifts a vector of dates by a period, while respecting business calendar rules.

cal_shift_end_of_month() shifts a vector of dates to the last day in the business month.

cal_shift(x, period = "1 day", convention = conventions$following,
  cal = calendar())

cal_shift_end_of_month(x, cal = calendar())

Arguments

x
[Date]

The dates to shift.

period
[character(1), lubridate::Period]

The period to shift x with. This is allowed to be a character such as "1 day", which is parsed by lubridate::period(), or a lubridate Period object generated from one of the helpers, such as lubridate::days().

The period is restricted to any combination of years, months, weeks, or days. Higher resolution periods such as hours, minutes, and seconds are not allowed.

Negative period values are allowed, to shift backwards.

convention
[character(1)]

A business day convention to follow if the date you land on after shifting is a holiday. These conventions only apply for monthly / yearly periods. For daily periods, the date is always shifted to the next available business day. The following conventions are available in the conventions object:

"following"

Choose the first business day after the given holiday.

"modified_following"

Choose the first business day after the given holiday unless it belongs to a different month, in which case choose the first business day before the holiday.

"preceding"

Choose the first business day before the given holiday.

"modified_preceding"

Choose the first business day before the given holiday unless it belongs to a different month, in which case choose the first business day after the holiday.

"unadjusted"

No adjustment is made.

"half_month_modified_following"

Choose the first business day after the given holiday unless that day crosses the mid-month (15th) or the end of month, in which case choose the first business day before the holiday.

"nearest"

Choose the nearest business day to the given holiday. If both the preceding and following business days are equally far away, default to following business day.

cal
[calendar]

A calendar.

Examples

library(lubridate)

#> 
#> Attaching package: ‘lubridate’

#> The following object is masked from ‘package:base’:
#> 
#>     date

x <- as.Date("2018-12-18") + 0:3
x

#> [1] "2018-12-18" "2018-12-19" "2018-12-20" "2018-12-21"

# The 22nd and 23rd are weekends so
# "2018-12-21" is shifted to the next available
# business day, "2018-12-24"
cal_shift(x, "1 day")

#> [1] "2018-12-19" "2018-12-20" "2018-12-21" "2018-12-24"

# The 25th is Christmas
# The following occurs because we skip over the weekend:
# "2018-12-20" -> "2018-12-24"
# Here, first we land at "2018-12-24" after skipping over the weekend,
# then we go one more business day forward to "2018-12-26", skipping
# over Christmas
# "2018-12-21" -> "2018-12-26"
cal_shift(x, "2 day")

#> [1] "2018-12-20" "2018-12-21" "2018-12-24" "2018-12-26"

# We could define a calendar where Christmas for that year is not a holiday
cal <- calendar()
cal <- holidays_remove(cal, "2018-12-25")
cal_shift(x, "2 day", cal = cal)

#> [1] "2018-12-20" "2018-12-21" "2018-12-24" "2018-12-25"

# ---------------------------------------------------------------------------
# Conventions

# In this case, moving forward 1 month would land on New Years. The
# default convention is to move to the "following" business day
cal_shift("2019-12-01", months(1))

#> [1] "2020-01-02"

# We can adjust this with the convention argument. Here we ask to move
# to the first business day before the holiday that we landed on
cal_shift("2019-12-01", months(1), "preceding")

#> [1] "2019-12-31"

# If you are okay with landing on New Years, use unadjusted.
cal_shift("2019-12-01", months(1), "unadjusted")

#> [1] "2020-01-01"

# 2010 was special because New Year was celebrated on the
# last day of December.
cal_shift("2009-12-31", years(1))

#> [1] "2011-01-03"

# In this case, "modified_following" can be useful. It generally
# shifts forward unless the holiday you land on is the end of the month,
# in which case it rolls back to the previous business day.
cal_shift("2009-12-31", years(1), "modified_following")

#> [1] "2010-12-30"

# Which means it gives the same result as preceding here
cal_shift("2009-12-31", years(1), "preceding")

#> [1] "2010-12-30"

# Shifting to the end of the business month
cal_shift_end_of_month("2019-03-14")

#> [1] "2019-03-29"

# Note that this can shift dates backwards if they are past the end of the
# business month, but not yet into the next month
cal_shift_end_of_month("2019-03-31")

#> [1] "2019-03-29"