Construct a calendar

These are various functions revolving around creating calendar objects, and setting the default calendar.

calendar() constructs a new calendar object. There are a large assortment of existing calendars to pick from, each with their own pre-existing holiday list.

empty_calendar() constructs a special calendar that has no predefined holidays, and no predefined business week. You can set the weekends on creation.

calendars lists the set of possible calendars that can be used.

calendar(name = calendars$united_states)

empty_calendar(weekends = c("Saturday", "Sunday"))

calendars

Arguments

name
[character(1)]

The base calendar to create. This should be one of the options listed in calendars.

weekends
[character]

Weekdays that should be special-cased as weekends. Weekends are considered "holidays," and are skipped over when shifting dates forward in the business week. Choose from:

"Sunday"

"Monday"

"Tuesday"

"Wednesday"

"Thursday"

"Friday"

"Saturday"

Format

An object of class vctrs_list_of (inherits from vctrs_vctr) of length 4.

Value

calendar() returns a new calendar.

empty_calendar() returns a new empty calendar.

calendars returns a named list_of<character>.

Examples

calendar()

#> <Calendar: united_states>

calendar(calendars$argentina)

#> <Calendar: argentina>

empty_calendar()

#> <Calendar>
#> Weekends: 
#>     - Sunday
#>     - Saturday

empty_calendar(weekends = "Monday")

#> <Calendar>
#> Weekends: 
#>     - Monday

# All available calendars
calendars

#> <list_of<character>[4]>
#> $argentina
#> [1] "argentina"
#> 
#> $argentina_merval
#> [1] "argentina_merval"
#> 
#> $united_states
#> [1] "united_states"
#> 
#> $united_states_settlement
#> [1] "united_states_settlement"
#>