Holidays

The following generate schedule objects corresponding to commonly used holidays.

hldy_washington_birthday(since = "1970-01-01")

hldy_memorial_day(since = "1970-01-01")

hldy_labor_day(since = "1970-01-01")

hldy_columbus_day(since = "1970-01-01", until = NULL)

hldy_veterans_day(since = "1970-01-01")

hldy_new_years_day(since = "1970-01-01")

hldy_martin_luther_king_jr_day(since = "1970-01-01")

hldy_independence_day(since = "1970-01-01")

hldy_thanksgiving(since = "1970-01-01")

hldy_christmas(since = "1970-01-01")

hldy_easter(since = "1970-01-01", until = "2040-01-01")

hldy_good_friday(since = "1970-01-01", until = "2040-01-01")

hldy_easter_monday(since = "1970-01-01", until = "2040-01-01")

hldy_july_fourth(since = "1970-01-01")

hldy_lincoln_birthday(since = "1970-01-01", until = "2040-01-01")

hldy_election_day(since = "1970-01-01", until = "2040-01-01")

Arguments

since
[Date(1)]

The initial date to start the schedule from.

until
[NULL / Date(1)]

An optional end date for the holiday schedule. Useful when constructing larger calendars, such as one for the NYSE, where holidays are occasionally only celebrated for a set number of years.

Examples

on_thanksgiving <- hldy_thanksgiving()

alma_search("2000-01-01", "2019-01-01", on_thanksgiving)

#>  [1] "2000-11-23" "2001-11-22" "2002-11-28" "2003-11-27" "2004-11-25"
#>  [6] "2005-11-24" "2006-11-23" "2007-11-22" "2008-11-27" "2009-11-26"
#> [11] "2010-11-25" "2011-11-24" "2012-11-22" "2013-11-28" "2014-11-27"
#> [16] "2015-11-26" "2016-11-24" "2017-11-23" "2018-11-22"