Add and remove holidays

holidays_add() adds new holidays to a calendar. Existing holidays are ignored. If a holiday you are trying to add has previously been removed, then it will be re-added.

holidays_remove() remove holidays from a calendar. These can be manually added holidays from using holidays_add(), or pre-existing holidays in that calendar (such as July 4th for a certain year in the "united_states" calendar). Dates that are not holidays are ignored.

holidays_added() lists the manually added holidays in a calendar.

holidays_removed() lists the manually removed holidays in a calendar.

holidays_add(cal, holidays)

holidays_remove(cal, holidays)

holidays_added(cal)

holidays_removed(cal)

Arguments

cal
[calendar]

A calendar.

holidays
[Date]

A vector of dates to add or remove from a calendar.

Examples

cal <- calendar()

date <- as.Date("2019-01-03")

# Notice that this day is not a holiday
cal_is_holiday(date, cal)

#> [1] FALSE

cal_shift("2019-01-02", "1 day", cal = cal)

#> [1] "2019-01-03"

# Now let's add it as a holiday for this calendar
cal_with_holiday <- holidays_add(cal, date)
cal_with_holiday

#> <Calendar: united_states>
#> Added holidays: 
#>    -  2019-01-03

# Now it registers as a holiday
cal_is_holiday(date, cal_with_holiday)

#> [1] TRUE

cal_shift("2019-01-02", "1 day", cal = cal_with_holiday)

#> [1] "2019-01-04"

# You can remove it afterwards
cal_back_to_normal <- holidays_remove(cal_with_holiday, date)
cal_back_to_normal

#> <Calendar: united_states>

# Dates that are already holidays are ignored if you try to add them
# This is New Years
cal_is_holiday("2019-01-01", cal)

#> [1] TRUE

holidays_add(cal, "2019-01-01")

#> <Calendar: united_states>

# If you don't want New Years to be a holiday in the US calendar, you
# can remove it
cal_no_new_years <- holidays_remove(cal, "2019-01-01")
cal_no_new_years

#> <Calendar: united_states>
#> Removed holidays: 
#>    -  2019-01-01

# Which means that this is an allowed date when shifting
cal_shift("2018-12-31", "1 day", cal = cal)

#> [1] "2019-01-02"

cal_shift("2018-12-31", "1 day", cal = cal_no_new_years)

#> [1] "2019-01-01"

# List the added holidays
holidays_added(cal_with_holiday)

#> [1] "2019-01-03"

# List the removed holidays
holidays_removed(cal_no_new_years)

#> [1] "2019-01-01"