Create a recurrence rule

These functions allow you to create a recurrence rule with a specified frequency. They are the base elements for all recurrence rules. To add to them, use one of the recur_*() functions.

daily() Recur on a daily frequency.

weekly() Recur on a weekly frequency.

monthly() Recur on a monthly frequency.

yearly() Recur on a yearly frequency.

daily(since = "1970-01-01")

weekly(since = "1970-01-01")

monthly(since = "1970-01-01")

yearly(since = "1970-01-01")

Arguments

since
[Date(1)]

The initial date to anchor the recurrence from. Depending on the final recurrence rule, pieces of information from this anchor date might be used to generate the event dates.

Details

The since argument is very important for both practical use, and speed. The default is set to the Unix epoch time, but there is no hard and fast rule for doing this. Remember that this is the first possible event date, so you may need to move this date backwards in time if you need to generate dates before 1970-01-01.

In terms of speed, it is more efficient if you adjust the since date to be closer to the first date in the sequence of dates that you are working with. For example, if you are working with dates in the range of 2019 and forward, adjust the since date to be 2019-01-01 for a significant speed boost.

As the anchor date, recurrences are often calculated relative to this date. As an example, a recurrence of "on Monday, every other week" would use the since date to find the first Monday to start the recurrence from.

There is no quarterly() recurrence frequency, but this can be accomplished with monthly() %>% recur_on_interval(3). The month to start the quarterly interval from will be pulled from the since date inside monthly(). The default will use a quarterly schedule starting in January since the default since date is 1970-01-01. See the examples.

Examples

rrule <- monthly() %>% recur_on_mday(25)

sch_seq("1970-01-01", "1971-01-01", rrule)

#>  [1] "1970-01-25" "1970-02-25" "1970-03-25" "1970-04-25" "1970-05-25"
#>  [6] "1970-06-25" "1970-07-25" "1970-08-25" "1970-09-25" "1970-10-25"
#> [11] "1970-11-25" "1970-12-25"

# Notice that dates before 1970-01-01 are never generated with the defaults!
sch_seq("1969-01-01", "1970-01-01", rrule)

#> Date of length 0

# Adjust the `since` date to get access to these dates
rrule_pre_1970 <- monthly(since = "1969-01-01") %>% recur_on_mday(25)
sch_seq("1969-01-01", "1970-01-01", rrule_pre_1970)

#>  [1] "1969-01-25" "1969-02-25" "1969-03-25" "1969-04-25" "1969-05-25"
#>  [6] "1969-06-25" "1969-07-25" "1969-08-25" "1969-09-25" "1969-10-25"
#> [11] "1969-11-25" "1969-12-25"

# A quarterly recurrence rule can be built from
# `monthly()` and `recur_on_interval()`
on_first_of_the_quarter <- monthly() %>%
  recur_on_interval(3) %>%
  recur_on_mday(1)

sch_seq("1999-01-01", "2000-04-01", on_first_of_the_quarter)

#> [1] "1999-01-01" "1999-04-01" "1999-07-01" "1999-10-01" "2000-01-01"
#> [6] "2000-04-01"

# Alter the starting quarter by altering the `since` date
on_first_of_the_quarter_starting_in_feb <- monthly(since = "1998-02-01") %>%
  recur_on_interval(3) %>%
  recur_on_mday(1)

sch_seq("1999-01-01", "2000-04-01", on_first_of_the_quarter_starting_in_feb)

#> [1] "1999-02-01" "1999-05-01" "1999-08-01" "1999-11-01" "2000-02-01"