Add to a schedule

sch_add_rrule() adds a rrule to a schedule.

sch_add_rdate() adds a rdate to a schedule. rdates are singular special cased dates that are forcibly included in the schedule.

sch_add_exdate() adds a exdate to a schedule. exdates are singular special cased dates that are forcibly excluded from the schedule.

sch_add_schedule() merges two schedules together.

sch_add_rrule(x, rrule)

sch_add_rdate(x, rdate)

sch_add_exdate(x, exdate)

sch_add_schedule(x, schedule)

Arguments

x
[schedule]

A schedule.

rrule
[rrule]

An rrule to add to the schedule.

rdate
[Date]

Dates to forcibly include in the schedule.

exdate
[Date]

Dates to forcibly exclude from the schedule.

schedule
[schedule]

A schedule to merge with x.

Details

In terms of priority:

An exdate will never be included.

A rdate will always be included if it is not also an exdate.

An event generated from a rrule will always be included if it is not also an exdate.

Combining two rrules into the same schedule is a way of joining them using an "or" condition. For example, joining monthly() %>% on_mday(5) with weekly() %>% on_wday("Tuesday") would translate to "the 5th day of the month or any Tuesday".

Examples

library(lubridate, warn.conflicts = FALSE)

on_easter <- yearly() %>%
  recur_on_easter()

on_christmas <- yearly() %>%
  recur_on_mday(25) %>%
  recur_on_ymonth("Dec")

on_labor_day <- monthly() %>%
  recur_on_ymonth("Sep") %>%
  recur_on_wday("Mon", 1)

sch <- schedule() %>%
  sch_add_rrule(on_easter) %>%
  sch_add_rrule(on_christmas) %>%
  sch_add_rrule(on_labor_day)

# Christmas, Easter, or Labor Day
sch_seq("2019-01-01", "2021-01-01", sch)

#> [1] "2019-04-21" "2019-09-02" "2019-12-25" "2020-04-12" "2020-09-07"
#> [6] "2020-12-25"

# Except Labor Day in 2019
sch2 <- sch %>%
  sch_add_exdate("2019-09-02")

sch_seq("2019-01-01", "2021-01-01", sch2)

#> [1] "2019-04-21" "2019-12-25" "2020-04-12" "2020-09-07" "2020-12-25"